Remove autoresizing mask from the title bar view.
The autoresize mask seems to be causing some views to get into a constraint layout loop in Mountain Lion. INAppStoreWindow is actually attached to the "theme view", a superview of the window's contentView. According to the docs contentView is the highest level accessible view, so I bet that they changed some behavior in there that's causing this issue. Instead we forgo autoresizing and make sure to recalculate the view frame manually whenever the window resizes.