Use CoreUI to draw the title bar and separator if available.
This uses private APIs and so is not included by default, but for
developers who are not releasing their apps in the App Store (or have a
separate non App Store build), the CoreUIRendering category on
INTitlebarView can be added to the project, which will automatically
load CoreUI and use it for rendering. This will draw 100% native title
bars with graceful fallback to custom drawing if the CoreUI drawing
function could not be loaded.

The preprocessor symbol INAPPSTOREWINDOW_NO_COREUI can be set to 1 to
disable CoreUI rendering; this is a convenience because Xcode's build
system cannot conditionally include a file in a target, but preprocessor
defines can be added for different build configurations.